6.3 Backport from hirsute

+++ Programs/timing.c
@@ -2,7 +2,7 @@
  * BRLTTY - A background process providing access to the console screen (when in
  *          text mode) for a blind person using a refreshable braille display.
  *
- * Copyright (C) 1995-2019 by The BRLTTY Developers.
+ * Copyright (C) 1995-2021 by The BRLTTY Developers.
  *
  * BRLTTY comes with ABSOLUTELY NO WARRANTY.
  *
@@ -256,7 +256,7 @@ millisecondsBetween (const TimeValue *from, const TimeValue *to) {
   };
 
   normalizeTimeValue(&elapsed);
-  return (elapsed.seconds * MSECS_PER_SEC)
+  return ((long int)elapsed.seconds * MSECS_PER_SEC)
        + (elapsed.nanoseconds / NSECS_PER_MSEC);
 }
